スケジュール管理のチームメンバーの表示名変更機能の実装（テストは未実装）
