--- app/helpers/team_members_helper.rb
@@ -1,2 +1,0 @@
-module TeamMembersHelper
-end
