+++ app/views/events/show.html.erb
@@ -10,7 +10,21 @@
         <th>△</th>
         <th>×</th>
         <% @team.users.sort_by(&:id).each do |member| %>
-          <th nowrap><%= member.name %></th>
+					<% if @current_team.admin_user_id == @current_user.id || member.id == current_user.id %>
+						<% team_member = TeamMember.find_by(user_id: member.id, team_id: @team.id) %>
+						<% if team_member.display_name.present? %>
+							<th nowrap><%= link_to team_member.display_name, edit_team_team_member_path(team_member) %></th>
+						<% else %>
+							<th nowrap><%= link_to member.name, edit_team_team_member_path(team_member) %></th>
+						<% end %>
+					<% else %>
+						<% team_member = TeamMember.find_by(user_id: member.id, team_id: @team.id) %>
+						<% if team_member.display_name.present? %>
+							<th nowrap><%=team_member.display_name %></th>
+						<% else %>
+          		<th nowrap><%= member.name %></th>
+						<% end %>
+					<% end %>
         <% end %>
       </tr>
       <% @events.sort_by(&:day_time).each do |event| %>
