+++ app/views/team/team_members/_display_name_form.html.erb
@@ -1,0 +1,6 @@
+<%= form_with(model: @team_member,url: team_team_member_path, local:true) do |f|%>
+  <%= f.label :display_name, '表示名' %>
+  <%= f.text_field :display_name %>
+  <%= render 'shared/form_error_messages', object: f.object, attribute: :display_name %>
+  <%= f.submit yield(:button_text), class: "form-btn btn-orange" %>
+<% end %>
